Create Naive Approach to check Prime number in C
We check the divisibility of the given number N with all the numbers starting from 2 to (N/2).
If it is completely divisible by any number i.e. remainder is zero after division, then the number is not a prime number.
If it is not completely divisible by a number between 2 and (N/2), then the number is a prime number.